Fix unwanted changes due to merge conflict
Unwanted changes to this branch were added from a merge conflict with
the main branch. These changes were needed to resolve the conflict.

+++ src/main/java/j_parser/utils/enums/AccessFlags.java
@@ -1,19 +1,19 @@
 package j_parser.utils.enums;
 
 public enum AccessFlags{
-    ACC_PUBLIC((short)0x0001),
-    ACC_FINAL((short)0x0010),
-    ACC_SUPER((short)0x0020),
-    ACC_INTERFACE((short)0x0200),
-    ACC_ABSTRACT((short)0x0400),
-    ACC_SYNTHETIC((short)0x1000),
-    ACC_ANNOTATION((short)0x2000),
-    ACC_ENUM((short)0x4000),
-    ACC_MODULE((short)0x8000);
+    ACC_PUBLIC(0x0001),
+    ACC_FINAL(0x0010),
+    ACC_SUPER(0x0020),
+    ACC_INTERFACE(0x0200),
+    ACC_ABSTRACT(0x0400),
+    ACC_SYNTHETIC(0x1000),
+    ACC_ANNOTATION(0x2000),
+    ACC_ENUM(0x4000),
+    ACC_MODULE(0x8000);
 
-    public final short FLAG;
+    public final int FLAG;
 
-    AccessFlags(short flag){
+    AccessFlags(int flag){
         this.FLAG = flag;
     }
 }
